Add build stamp

+++ library_builders.sh
@@ -430,6 +430,7 @@ function build_fftw {
 }
 
 function build_cfitsio {
+    if [ -e cfitsio-stamp ]; then return; fi
     if [ -n "$IS_OSX" ]; then
         brew install cfitsio
     else
@@ -440,4 +441,5 @@ function build_cfitsio {
             && ./configure --prefix=$BUILD_PREFIX \
             && make shared && make install)
     fi
+    touch cfitsio-stamp
 }
